Refactor : string, number 나누지 않고 BaseballConstant로 묶어서 관리, 인스턴스 생성 못하도록 private 생성자 추가
