+++ src/main/java/constant/BaseballConstant.java
@@ -1,7 +1,13 @@
 package constant;
 
-public class StringConstant {
+public class BaseballConstant {
 
+    private BaseballConstant() {
+    }
+
+    public static final int START = 1;
+    public static final int END = 2;
+    public static final int MAX_LENGTH = 3;
     public static final String BALL = "볼";
     public static final String STRIKE = "스트라이크";
     public static final String NOTHING = "낫싱";
