--- src/main/java/constant/NumberConstant.java
@@ -1,9 +1,0 @@
-package constant;
-
-public class NumberConstant {
-
-    public static final int MAX_LENGTH = 3;
-    public static final int START = 1;
-    public static final int END = 2;
-
-}
